cmake/modules/GenerateLuaFiles: remove unneeded cmake call
The C file is already generated at configure time so it will be
available at build time.
In addition to be unneeded, this call could be an issue if `cmake` is
not the right cmake executable to call, and if we tried to build the
project within a directory outside of the source directory.